Rename sensing test files, should be other way around!
 - Whoops, my bad
